requirements.txt
pandas==2.2.2
numpy==2.0.2
scikit-learn==1.5.2
scipy==1.13.1
PyYAML==6.0.1